requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
django-cors-headers
djangorestframework==3.14.0
docopt==0.6.2
gunicorn==21.2.0
idna==3.6
joblib==1.3.2
numpy
packaging==23.2
pipreqs==0.4.13
pytz==2023.3.post1
requests==2.31.0
scikit-learn
scipy
sqlparse==0.4.4
threadpoolctl
tzdata==2023.3
urllib3
yarg==0.1.9
pymongo
bcrypt